refactor: move util to new promql utils
