+++ src/WingmanDataTrail/ListControls/MetricsSorter/fetchers/fetchDashboardMetrics.ts
@@ -2,10 +2,9 @@ import { getBackendSrv, type BackendSrvRequest } from '@grafana/runtime';
 import { type Dashboard, type Panel } from '@grafana/schema';
 import { limitFunction } from 'p-limit';
 
-import { logger } from 'tracking/logger/logger';
-import { isPrometheusDataSource } from 'utils/utils.datasource';
-
-import { extractMetricNames } from './extractMetricNames';
+import { logger } from '../../../../tracking/logger/logger';
+import { isPrometheusDataSource } from '../../../../utils/utils.datasource';
+import { extractMetricNames } from '../../../../utils/utils.promql';
 
 interface DashboardSearchItem {
   id: number;
